Fixed string pool issue with the main toolbar.
Previously, the string index for a particular button was calculated by
subtracting the ID of the first button from the ID of that button.
However, if there were any gaps between the button IDs, the string index
would be wrong.

This commit updates the main toolbar code so that the mapping between
button IDs and strings is stored. This fixes #49.