Added shell extension blacklist
Some shell extensions are known to be incompatible with Explorer++. For
example, the OneDrive shell extension present in Windows 10 causes
Explorer++ to crash when trying to display the background context menu
(for details see #35).

As finding the exact cause of issues like that can be non-trivial (it
could be due to a bug in Explorer++, or a bug in the extension, or an
issue with how the two interact), some basic blacklisting functionality
has been added.

This functionality uses a hardcoded list of shell extension CLSIDs to
detect and prevent extensions that are known to incompatible from being
loaded. The version of Windows is not taken into account, so an
extension present in multiple versions of Windows will be blocked on all
of them. Currently, the only extension on the list is the OneDrive file
sync extension.

While the best solution in the long term might be to find out precisely
why there is an issue, this workaround fixes #35 for now.