Fixed issue with file creation on Windows 10
Previously, when creating a file via the new menu in Windows 10, the
file would not be put into edit mode. This seems to have been caused by
the fact that an undocumented flag value is passed to
SelectAndEditItem() on Windows 10. As Explorer++ didn't handle this
value, the rename operation never took place.

The code has been updated so that this value is now handled. This fixes
#48.