Updated code so that rename operations are silent
Previously, rename operations could display a progress dialog box.
However, this meant that if you put a file into edit mode, changed the
name, then pressed tab (while still in edit mode), the next file would
be selected, but would not put into edit mode. This was because the
progress dialog was stealing focus. While this doesn't seem to have been
a problem on Windows 7, it was a problem on at least Windows 10.

Now, renames are always performed silently. This means that tabbing to
another file works correctly and fixes #50. Also, as renaming a single
file or group of files should be fairly quick anyway, a progress dialog
shouldn't really be required.