Updated main solution to include translations.
Each of the translation DLLs is now built as part of the main solution
when in Release mode. This has the advantage that the executable and
DLLs are built at the same time (and therefore don't need to be built
independently). This makes it easier to manage the build process.

Note that as the platform doesn't matter for resource DLLs, both the
Win32 and x64 builds target the Win32 platform. Ultimately, building the
translations twice (once during the Win32 build and once during the x64
build) is unnecessary, but it simplifies the build process, since there
doesn't have to be special handling for one of the configurations.

Also, as part of this change, stub resource.h files have been added to
the translation projects. Using these files means that (1) the .rc files
don't have to be updated (which is important, as the translated .rc
files should be the same as the English .rc file) and (2) it's easier to
update the actual resource.h file that's used.