Fix issue with taskbar thumbnail generation
When generating taskbar thumbnails for the tabs (on Windows 7 and
later), the application creates an image designed to fit within the
bounds provided by WM_DWMSENDICONICTHUMBNAIL.

However, because the target width/height were being calculated using
integers, the results weren't correct and this would sometimes cause the
width/height of the final image to be greater than the maximum
width/height allowed. This then meant that the images wouldn't be
displayed.

To resolve this, the width/height calculation is now done using
floating-point arithmetic. This fixes #55.