Include several different sizes in main icon
Previously, the main icon contained a single 48x48 image. Now, it
contains the following image sizes:

- 48x48
- 32x32
- 16x16

This allows other applications (e.g. the volume mixer) to retrieve a
different size of the image if necessary. Note that as part of this
change, the small icon has been removed, as it is no longer necessary.
This fixes #24.