Replace Pantheios with Boost.Log
Pantheios and STLSoft aren't maintained any more. That made it difficult
to use newer versions of Visual Studio to compile the code. To alleviate
this, the Boost.Log library is now used to perform logging. This also
has the advantage of reducing the number of dependencies. Additionally,
since Boost is included via NuGet, there shouldn't be a need to check
out dependencies manually any more.

Note that warning C4714 (function 'function' marked as __forceinline not
inlined) is disabled in each project, as the warning is triggered when
using the Boost.Log library.