Only process WM_SIZE once fully initialized
WM_SIZE can be sent as a result of actions taken in WM_CREATE. This was
an issue, as it meant that windows that were assumed to exist didn't
(they hadn't been created yet), ultimately resulting in reading from
uninitialized variables. To resolve this, WM_SIZE will only be processed
now once WM_CREATE has been handled.

This issue was found with Dr. Memory.