Update font used by tab control
Previously, the font used by the font control was a fixed size. This
could cause issues on systems using a different DPI (e.g. the text would
be smaller than other standard text). Using a standard font ensures that
the font will scale based on the display settings set.

Note that because the tab control is currently a fixed size, on systems
where the font is larger (e.g. because the DPI is higher), the text will
be cut off. However, this change forms at least part of the work needed
to support higher DPI displays.