Delete items from listview before freeing internal data structures
Previously, when navigating to a folder, the internal data structures
associated with the last directory would be freed and then the items in
the listview would be deleted. This, however, could cause issues in some
cases, as an NM_CUSTOMDRAW message could be processed in between the
time when the data structures were freed and the time when the listview
items were deleted.

Going through the following steps would occasionally trigger the bug:

1. Navigate to a folder that has several items.
2. Switch to a DVD drive with no disc in by clicking on the associated
   button in the drives toolbar.
3. The application may crash due to trying to access memory that's been
   freed.

Note that going through the above steps while running the application
under Dr. Memory would trigger the bug much more reliably.

When this bug was triggered, the following calls would occur:

1. BrowseFolder()
1.1. ResetFolderMemoryAllocations()
1.2. BrowseVirtualFolder()
1.3. pShellFolder->EnumObjects()
2. OnCustomDraw() (called in response to an NM_CUSTOMDRAW message)
2.1. m_pActiveShellBrowser->QueryFullItemName()
2.2. QueryFullItemNameInternal()
2.3. ILCombine() (this is the call that would ultimately cause problems,
     due to it trying to access freed memory)
1.4. ListView_DeleteAllItems()

To try and resolve this issue, the listview items are now deleted before
the internal data structures are freed. This bug may have been the cause
of #31.