Merge ShellBrowser project into main project
There are cases where the ShellBrowser displays strings that need to be
translated. To allow this to happen, the strings first need to be placed
in a resource file. However, there's not really any way of including
resources in a static library, so including the files within the main
project means they can use the resources set up there.