Fix treeview notify handler
Previously, the treeview control always returned 0 in response to a
WM_NOTIFY message. For some notifications, this might have been ok (e.g.
if the return value was ignored), but it could have caused issues with
other notifications (e.g. if 0 was supposed to mean that the application
had handled the message).

One visible effect this had is that it prevented treeview labels from
being unfolded. That is, if you moused over a treeview label that was
only partially visible, no tooltip would be shown. This was probably
because the return value of 0 had a specific meaning for one of the
WM_NOTIFY messages being sent.

To fix this, the treeview control now calls DefSubclassProc() and
returns the value from that.