Remove VS2012 and VS2013 platform toolset declarations
The projects in the solution may still compile with those versions of
Visual Studio, but it wouldn't work out of the box (as the Boost
libraries that are imported via NuGet specifically target VS2015).
Additionally, I have no simple way of checking whether the projects
actually do compile in those versions.