Convert user documentation to RST format
The docs are compiled using Sphinx. Converting them to RST allows them
to be displayed online at Read the Docs (https://readthedocs.org/).

Note that one piece of functionality not imported from the original HTML
Help files is the ability to click on certain parts of an image (which
would then take you to the documentation for the particular element you
clicked on). If necessary, this functionality could potentially be added
in the future.

Additionally, some of the pages (e.g. the regular expression pages and
appendices) have not been converted at this point.