Update way in which column data is retrieved and set
Previously, column data was retrieved on a background thread. That
thread would iterate through each item in the listview, retrieve the
column data then insert it into the listview.

Now, the column data is still retrieved in a background thread, but
items are only queued to the thread as necessary (e.g. as items are
scrolled into view). Additionally, the background thread no longer
interacts with the listview at all.

The updated implementation has a number of advantages:

- As mentioned above, items are only queued when necessary. This means
  that not all the results have to be loaded at once.
- Because the background thread no longer touches the UI, it's
  potentially easier to lock its access to shared data. Previously,
  acquiring a lock in the background thread and then attempting to lock
  the main thread could result in a deadlock. The background thread
  would try to update the UI, which would require work to be done on the
  main thread, but the main thread would be locked waiting on the
  background thread.
  Note that this locking wasn't actually added. It wouldn't have been
  safe to do so, for the above reason.
- Since the background thread no longer touches the UI, it may help to
  cut down the crashes that can occur when navigating to a different
  folder while column results are being filled. Specifically, the first
  thing the previous background thread did for an item was to look that
  item up in the listview. That could fail (if a navigation had occurred
  since the item was queued), potentially causing a crash.
- Because the previous implementation updated the UI directly, it was
  possible for a UI update to be done after a navigation had occurred.
  That is, column data for an item in a previous folder would be
  inserted into the listview for the current folder.
  In the new implementation, this isn't possible. Results are posted
  back to the main thread and the main thread will ignore results for
  previous folders.
- Both the listview item index and column index are looked up at the
  point where the column data is set (in the main thread). Previously,
  the item index and column index were determined before the column text
  was retrieved. This meant it was possible for the indexes to change
  before the text had been set.
- Parts of the implementation are simplified, since a library is used
  for the thread pool functionality (rather than custom code).
- It's trivial to increase the number of threads used in the thread pool
  (though that may not be of too much use in this specific case, as the
  work is IO bound).

However, there is still at least one remaining issue. The background
thread shares state with the main thread (the same object is used for
both). At the moment, there's no locking that's done. That means that if
a navigation occurs while column data is being retrieved, or a file is
modified or removed, the application could still crash.

Adding locks to the appropriate code may be easier now that the
background thread doesn't touch the UI, but ultimately, there's still
too much state that's being shared.

Another potential issue with the new implementation is that the library
being used for the thread pool functionality
(https://github.com/vit-vit/CTPL) hasn't been updated since 2015. It may
not receive any future bug fixes. The library is fairly simple though
and could potentially be replaced if necessary.

Any replacement would need the following functionality:

- The ability to control the number of threads in the pool.
- The ability to push tasks to a queue.
- The ability to clear the queue.
- The ability to retrieve results using std::future.