Update folder size threading code
Previously, folder sizes were calculated in an independent background
thread. This had a lot of the same disadvantages as the background
thread that retrieved and set column text (e.g. the UI was updated
directly from the background thread).

This code has now been updated so that folder sizes are simply
calculated as part of the normal process to retrieve column text. This
simplifies the code and brings the same advantages as the change to
update the column threading model (see
75440e01635e7b749682f08cf13310a4d4689007).

One potential disadvantage of this approach is that it means that folder
sizes are calculated along with the rest of the column items. If a
particular folder size takes a long time to calculate, other column
details won't be displayed. If that turns out to be too big of a
problem, the size of the thread pool could potentially be increased.