Start moving column retrieval functions out of CShellBrowser class
Having the functions outside of the class makes it much easier to
guarantee they aren't using any data that's part of the class (important
when the functions may run on a background thread). Note that not all
the functions have been moved yet - those that haven't been moved are
using preference values stored within the class.

Also note that, at the moment, the functions that have been moved are
still using the ItemInfo_t struct. This struct provides more information
than is necessary. It will be updated at some point so that only the
necessary fields are shared.