Pass copies of data to column retrieval function
With this change, the column retrieval code should now be entirely
thread safe. All input data is copied, rather than looked up. This means
that whether a column retrieval process finishes before a navigation or
after makes no difference. The data that the thread holds won't be
changed or deleted.

This should fix #68.