Start moving configuration variables into separate data structure
Currently, configuration values are largely stored within the
Explorerplusplus class. That has a number of disadvantages:

- It makes the class larger than it otherwise would be.
- It makes it hard to move components of the class into independent
  classes, as the components depend on the configuration (so there would
  need to be a lot of shared state passed around).

The second point is the more important at this stage. Breaking up the
Explorerplusplus would have a number of benefits. To do that properly,
the configuration needs to be handled in a better way.

Note that the changes started in this commit don't represent the final
solution at all. Currently, the aim is just to pull the variables out of
the Explorerplusplus class.