Embed Lua in application
The aim ultimately is to support plugins written in Lua. An advantage of
supporting plugins is that various improvements will have to be made to
the internals of Explorer++, to provide a simple external interface.
Obviously, plugins would also allow users to extend the application
without having to update the application itself.

Using Lua as an embedded language has a few advantages over other
alternatives:

- Lua is lightweight. The impact on compile time is low and the impact
  on executable size is small (it adds about 100 KB).
- Lua is simpler than a language like C++. Having C++ plugins would
  allow more flexibility in what the plugins are allowed to do, but C++
  is more difficult to write, the plugin system would likely be more
  complicated and it would be much harder to do safely. Using a simpler
  language makes it easier to write a plugin.
- Following on from the above point, it should be easier to have the
  plugin system be fairly safe. As the Lua code will ultimately be
  calling C++ code, it won't be completely safe (the C++ code could very
  well have bugs), but it should be harder for Lua plugins to affect the
  stability of the application itself.
- It's simple to allocate each plugin its own Lua state. This completely
  isolates each plugin. This is potentially more difficult to do with
  other languages.
- A general advantage of using an embedded language is that it's
  possible to create a basic interface that allows commands to be
  executed interactively. With a C++ API and plugins, this wouldn't
  really be possible.

Potential disadvantages:

- There are some things a Lua plugin wouldn't really be able to do, such
  as call the Windows API. Lua plugins also wouldn't be able to work
  with PIDLs (only stringified paths). That could make it harder for a
  plugin to request that a non-filesystem folder (e.g. the recycle bin)
  be opened. The application would probably need to provide an interface
  that allows non-filesystem folders to be queried/opened.

At this stage, the intention is to embed Lua, add a user interface that
allows scripting commands to be run interactively and start working on
API to provide to the Lua code.

Including the functionality on the master branch means that it won't
become out of sync with other changes and improvements made to the
internals to support plugins can be released immediately.

The functionality will be disabled by default, with a command line
switch required to enable it. Given how simple Lua is to embed, there
should be no real downside to embedding it in the application, even if
it's not actually used.

When the functionality is disabled, no Lua state will be created and no
Lua code will be run. This should limit he impact of any changes. Even
if the API provided to Lua is buggy, it won't actually be called
anywhere and no instance of Lua will be running.

Note that, currently, Lua is embedded as a static library. This means
that the application can still be run as a single executable. Switching
to a DLL would have two advantages:

- The lua NuGet package (https://www.nuget.org/packages/lua/) could be
  used. That package only provides DLLs.
- It would make debugging Lua scripts using something like MobDebug
  (https://github.com/pkulchenko/MobDebug) easier. While it is possible
  to debug with an embedded, static version of Lua, it's not quite as
  straightforward.

This closes #132.