Add basic tabs API
Currenty, this API has a single method: create (to create a new tab).
The API method can be called from Lua in the following way:

tabs.create("C:\\")

The API method doesn't return anything yet. There's also presently no
validation done on the argument that's passed in. Note that this almost
certainly doesn't represent the final form of the API, as it will likely
be updated to allow you to specify other options (e.g. the view mode,
sort mode, etc).

Additionally, the tabs table in Lua is explicitly set up as read-only,
to prevent Lua scripts from attempting to modify it.

This is part of the work for issue #133.