Add Lua scripting dialog
This dialog allows Lua commands to be run interactively.  The Lua state
is initialized in the same way it would be for plugins (i.e. the plugin
API methods are all binded in).

Currently, this functionality is disabled by default. The following
command line switch must be passed in to enable it:

explorer++.exe -enable_lua_scripting

Without this switch, the main menu item that opens this dialog will be
removed at startup.

This dialog will be used to do the initial testing of the plugins API.
There won't yet be any functionality to load or manage plugins.

This closes #134.