Add tabs API method to get details on a specific tab
Currently, this just returns a Tab userdata structure containing the tab
location, or nil if no tab with the specific ID is found. Note that a
__tostring method is set up for the Tab userdata. This allows certain
debugging/inspection modules in Lua (e.g. MobDebug) to display
information about the object (usually, they can't introspect userdata
types).

The new method can be called from Lua in the following way:

tab = tabs.get(0)

This is part of the work for issue #133.