Trigger event on tab creation
When a tab is created, an event is now triggered. The Boost Signals2
library is used to trigger the event and manage observers.

The intent is to add more of these sorts of events that plugins can
observe. This particular event is potentially useful for plugins, as
they can use it to be notified when a tab is created.