Load plugins on application startup
Plugins will be loaded from a "plugins" folder in the same directory as
the executable. Currently, there's no way to manage, enable, disable or
remove plugins. Once a plugin has loaded, it will stay loaded for the
lifetime of the application.

This functionality is behind a command line switch:

explorer++.exe -enable_plugins

Without this switch, no plugins will be loaded.

This is part of the work for #136.