Allow plugins to apply basic color changes to listview and treeview
The background color and text color can be set for each. Listview colors
will be applied to each existing listview, as well as new listviews
created during tab creation. The changes aren't persisted in any way,
but this doesn't matter too much, since a plugin in the plugins
directory will be loaded each time the application starts and can set
the colors then (which has the effect of persisting the changes).

Currently, there's no way to revert to the default colors.

The following Lua code can be used to set the colors in the listview and
treeview:

ui.setListViewColors("#1E1E1E", "#FFFFFF")
ui.setTreeViewColors("#1E1E1E", "#FFFFFF")

The first color is the background color, the second the text color.

This is part of the work for issue #137. It's also required to be able
to implement the changes for #115.