Add initial code to parse accelerator strings specified in plugin manifest
This is the first step towards allowing plugins to register specific
keyboard accelerators.

With this change, several test cases have also been added for the
parsing functionality. A test has also been added for the manifest
parsing.

Note that currently, the TestExplorer++ project is setup to use a static
library produced by the Explorer++ project. Typically, if you want to be
able to easily unit test some functions, you need those functions to be
in a static library that you can link into your test project. However,
the Explorer++ project produces an executable, rather than a static
library, which is an issue.

There's an alternative solution discussed in https://stackoverflow.com/a/19790019
and described in more detail at https://stackoverflow.com/a/19710979.
That solution is what's used here. The Explorer++ project produces both
an executable and a library composed of all objects. The test project
then links against that library. This isn't necessarily ideal, but
allows some initial unit tests to be added easily without having to turn
the main Explorer++ project into a static project.