Update listview file renaming code
Previously, file renaming in the listview would be blocked if the parent
folder was a virtual folder. Since items in virtual folders can now be
renamed, that check is no longer necessary. Instead, a check is done to
see if the selected file(s) can be renamed. If they can, the rename
process is initiated.