Ensure that items are put into rename mode when created
This was previously working, but was broken in one of the directory
monitoring updates. The code necessary to put items into rename mode
when created has been set back up.

This also fixes a few issues with the code:

- The PIDL of the new item was never being freed.
- The variable that kept track of whether an item had to be renamed
  wasn't being cleared when the directory changed, even though it's only
  ever relevant for the current directory.

This change also slightly simplifies the code, since it means that the
CShellBrowser class puts items into rename mode, rather than sending a
message back to the parent window for the Explorerplusplus class to
handle.