Allow plugins to register keyboard accelerators
Accelerator key combinations are specified in the manifest. When an
accelerator is invoked, the associated plugin will be notified via a
callback (the plugin will need to register this callback first).

At the moment, only new accelerator key combinations can be added. It's
not possible to change existing accelerators.

The sort_tabs_by_name plugin has been updated to illustrate how to
register and handle an accelerator combination.