Update tab structure passed to plugins
Settings specific to the view of the folder within the tab are now
contained within a FolderSettings struct inside the Tab struct. This
helps to separate settings related to the tab itself from settings
related to the view of the folder within the tab.

This is part of the work for issue #133.