Fire event when tab is locked/unlocked
This slightly simplifies the code, since it means that a tab can now be
locked/unlocked just by calling a method on it. Additionally, it means
that the code to update the tab in the user interface can be completely
separate from the code to update the lock status in the Tab class.
Finally, it also means that plugins could observe changes to the lock
status in future.