Update two tab retrieval methods so that they throw an exception on failure
Previously, GetTabByIndex() and GetTabIndex() could fail and return no
data. That would really only happen if there was a bug, though. Each tab
in the user interface should have an associated tab object and each
object should have an index. If either of these invariants are broken,
there's a bug.

Therefore, these functions now throw an exception when they fail. This
reduces the amount of error checking that needs to be done elsewhere
(e.g. to verify that the tab returned is valid) and simplifies parts of
the code. It also provides a clearer indication that failure represents
a hard error.