Update way in which tab settings are passed to CreateNewTab()
Previously, a pointer to a tab settings structure was passed. This
meant, however, that either all the settings in the structure had to be
specified at once or none of them. Now, the structure always has to be
passed in, but its individual members are all optional.

This is important because it will allow plugins to specify some values
without having to specify all values.