Allow plugins to specify some initial tabs settings
The following properties can now be set by a plugin when creating a tab:

- name
- locked
- addressLocked

This commit also changes the structure of the create call. Rather than
passing the location directly, it's now specified in the following way:

tabs.create({location = "C:\\"})

Additional properties can be added to the first argument. For example:

tabs.create({location = "C:\\", locked = true})

This is part of the work for issue #133.