Change how global folder settings are pushed to tabs
There are some settings that are global to the application and can't be
changed on a per-tab basis (e.g. whether or not to show file
extensions).

Previously, there were a few issues with the way these settings were
being pushed to individual tabs:

- When creating a tab, some of the settings would be set twice.
- Each tab would hold a copy of the settings, even though it wasn't
  possible to update the settings on a per-tab basis.
- Each tab would initialize its own copy of the settings in case they
  weren't supplied (meaning defaults were effectively being set in two
  places).
- Because each tab held its own copy of the settings, updating a setting
  would mean updating the value in each individual tab.
- It was possibly to query the value of one of the settings for an
  individual tab, even though the settings were global.

To fix these issues, a pointer to the global folder settings is supplied
once (during construction of a tab). Each tab then just holds a copy of
the pointer for its lifetime.