Use Better Enums for ViewMode enum
This should provide two distinct advantages:

1. Conversion to string will be done automatically (important when
   passing enum values to a plugin).
2. It'll be possible to safely convert integer values back into enum
   values.

As a side effect of this change, the ViewMode enum now acts more like an
enum class, meaning the enum values are scoped to the enum itself. Which
is positive, as it means the values are no longer global.

One weakness of the library is that enum values may need to prefixed
with a "+" when performing a comparison. This is considered to be an
acceptable trade-off, however, given the other advantages.

The version of Better Enums being used is 0.11.1.