Set default file and folder icons when browsing a folder
Previously, these were only being set once (in the constructor of the
CShellBrowser class). The issue with that was that if file associations
changed, the folder would be refreshed (the system image list is
refreshed, which requires all icons to be retrieved again), but the
default file and folder icons wouldn't be retrieved again, causing the
icons used to be incorrect.

Now, the default file and folder icons are retrieved every time a folder
is browsed. Which is inefficient (most of the time, the icons don't need
to be retrieved again), but fixes the above issue.