Adjust how OnTabChangeInternal() is called when closing a tab
Previously, this function was being called after the closed tab had been
removed from the window, but before it had been removed from the
internal set of tabs. This then meant that during the
OnTabChangeInternal() call, the number of tabs in the window would no
longer match the number of internal tabs (which is an invariant that's
supposed to always hold).

This was causing issues with the taskbar thumbnails code, as that code
now indirectly relies on the fact that the number of tabs held
internally is the same as the number of tabs in the window.

To fix that, OnTabChangeInternal() is now only called at the end of
CloseTab().