Update how tab tooltips are generated
Previously, tab tooltips were always set to the parsing path of the
folder within the tab. That didn't work so well for virtual folders like
"This PC" which would display a user-unfriendly GUID (as that's
ultimately what the parsing path for that folder is).

Now, tab tooltips are generated in the same way as the text shown in the
address bar. In cases where a GUID would be displayed, the friendly name
is used instead.