Retrieve file infotips in a background thread
Previously, file infotips were being retrieved on the main thread.
Because of this, the user interface could hang at times. This was
particularly bad, as all you would need to do is accidentally leave the
mouse on a large folder for a short amount of time and the entire
application would hang for multiple seconds.

To fix this, infotips are now retrieved on a background thread. Once the
infotip has been retrieved, LVM_SETINFOTIP is used to set the infotip
for the associated item. Note that this means that once an infotip has
been set, it won't be re-requested. To see an updated infotip, you would
need to refresh the folder. This is considered to be an acceptable
trade-off, given the advantages it brings.