Fix issue with status bar display at startup
UpdateStatusBarText() is called as the tabs are restored on startup. The
problem was that this was being done before the SetStatusBarParts()
function was called. This then meant that the calls to set the text in
parts 2 and 3 of the status bar would fail and they would be blank when
the application started.

To fix this, SetStatusBarParts() is now called as soon as the status bar
is created.