Automatically save settings every 30 seconds
This helps to prevent tabs and other settings from being lost if the
application crashes. The potential performance impact of this should be
very low. Saving the settings to the registry, for example, takes about
0.02 seconds in debug mode (though the exact time taken will obviously
vary based on the device). Therefore, it's unlikely that saving
periodically will have any observable impact on the user interface, even
though it's done on the main thread.