Share icon cache between tabs
This has several advantages:

1. If you close a tab, then open a new tab in the same folder a short
   time later, the icons are likely to be cached.
2. If you open a folder in one tab, then open it in another tab a short
   time later, the icons are likely to be cached.

One potential downside of this approach, as compared to the previous
approach, is that the size of the icon cache doesn't increase as the
number of tabs grows. Previously, each tab had its own icon cache, so
the total number of icons that could be cached would grow as the number
of tabs increased.

However, this shouldn't really be an issue. The maximum size of the
cache has been set to 1000 items, which should be more than enough.