Add VS 2019 PlatformToolset value
VS 2019 is binary compatible with VS 2017 (see https://devblogs.microsoft.com/cppblog/cpp-binary-compatibility-and-pain-free-upgrades-to-visual-studio-2019/),
which means that the code compiles and links cleanly under VS 2019, even
though some of the dependencies are built with VS 2017.

Note that the way in which PlatformToolset values are specified in the
vcxproj files has been tweaked. Using the previous format worked, but
resulted in the "General" configuration property page being partly blank
(e.g. the "Target Name" and "Target Extension" fields would be blank)
for projects created in newer versions of Visual Studio.