Update CloseTab() and RemoveTabFromControl()
There was previously an attempted fix for an issue that arose when
closing a tab (see 44603e69ec950e00b7357bc3d8faaa513bb434b3). That issue
was caused by the fact that OnTabSelectionChanged() was being called at
a point at which the number of tabs in the UI no longer matched the
number of tabs stored internally.

To try and fix that, the call to OnTabSelectionChanged() was moved to
the end of CloseTab(). However, that was causing another issue, outlined
below:

1. CloseTab() would call DestroyWindow(tab.listView).
2. The DestroyWindow() call would cause a WM_SETFOCUS message to be sent
   for the newly selected listview.
3. The handler for WM_SETFOCUS would call UpdateToolbarButtonStates(),
   to update the buttons in the main toolbar. This process relied on
   accessing the current tab. However, references to the current tab
   were still pointing to the tab that had just been destroyed, meaning
   this function would attempt to access a deleted object.

To fix the underlying issue, RemoveTabFromControl() has been updated so
that it first changes the tab selection before removing a tab. That's a
safe operation, as the number of tabs in the control will still match
the number of tabs stored internally. It then removes the target tab
from the UI. When CloseTab() calls DestroyWindow(tab.listView), the tab
selection has already changed, meaning references to the current tab are
valid. There is a small gap in between the time when a tab is removed
from the control and the time when it is removed from the set of
internal tabs, but there shouldn't be any code running during that time
that breaks because of that.

Finally, the constructor for the Tab class has been marked as explicit.
This is to prevent issues that might arise when updating a function that
takes an int tab ID to a function that takes a Tab reference. Without
using explicit, code passing an int will continue to compile even though
it's wrong.