Forward declare a number of items in the Explorer++.h header
This should help compile times in some cases, since it means that when a
header that used to be included in Explorer++.h changes, only a few
source files need to be recompiled, rather than every file that includes
Explorer++.h (of which there are currently about 30). It may also help
to reduce compile times in the general case, since the files that were
being included in Explorer++.h themselves included other files and had
other declarations.

Finally, a few unnecessary header includes have been removed.