Save tabs in correct order
One of the recent changes that's been made is to retrieve the list of
tabs from the internal set of tabs (rather than from the tab control
itself). This is fine in most cases, because what you need is the total
set of tabs and not their specific order.

However, when saving the list of tabs, you do need to retrieve them in
the order in which they're displayed in the UI, so that you can restore
them in that order. Therefore, a function has been added to retrieve the
list of tabs in order and this function is used when saving the tab
data.