Standardize the way folder columns are stored and propagated
Previously, folder columns were being stored and propagated in a few
different ways. In some cases, they were being stored as individual sets
of lists (one for each folder type). In other cases, the lists were
grouped together into a struct and the struct was passed around. In at
least one case, the lists were being passed as individual arguments to a
constructor.

This change standardizes the way folder columns are stored and passed
around between components. The columns are now always stored in a struct
and are passed around in that struct. Aside from being cleaner, this
also removes various pieces of duplicate code (in cases where each of
the lists was being specified individually).