Update how tab signals are managed
Previously, a signal object was defined for each type of tab event (e.g
tab create, tab update, etc) and a custom connection method was defined
for each signal.

This worked, but resulted in boilerplate code being added for every
signal. It also meant that you could only call overloads of the signal
connect() method in certain cases (when the custom connection method
passed the other arguments on).

Now, each signal is wrapped by a class and that wrapper class defines a
consistent connection method that simply forwards its arguments to the
underlying connect() method. This allows overloaded versions of the
connect() method to be called when necessary. It also cuts down the
amount of boilerplate that has to be written to add a new signal.