Allow accelerators to be remapped via plugin system
Via the manifest, a plugin can now remap accelerator keys. This is done
by providing a list of commands and the updated key bindings.

This type of configuration would probably be better stored along with
the rest of the configuration data (in the registry/XML file), but
implementation through the plugin system is easier, since no UI needs to
be built, reading the data is easier and there's no need to write the
data.