Fix WM_COMMAND handler for main window (closes #163)
Previously, it was possible for the handler to interpret control
notifications as menu/accelerator notifications. This is because the
wParam high word wasn't being properly checked.

In particular, it lead to an issue where if you put a file in a listview
into rename mode and typed a character, an EN_CHANGE notification would
be sent, but interpreted as TABTOOLBAR_CLOSE. This would result in the
current tab being closed (see issue #163).